Add partials to allow starring old messages
